requirements.txt
# This file is auto-generated from pyproject.toml by sync_requirements.sh
# DO NOT EDIT MANUALLY - Edit pyproject.toml instead and run: ./sync_requirements.sh
# Generated on: Sun Sep 21 10:49:52 EDT 2025
#
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
alembic==1.16.5
    # via mlflow
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.4
    # via fastmcp
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via
    #   google-auth
    #   mlflow-skinny
    #   mlflow-tracing
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   flask
    #   mlflow-skinny
    #   uvicorn
cloudpickle==3.1.1
    # via mlflow-skinny
contourpy==1.3.2
    # via matplotlib
cryptography==45.0.7
    # via
    #   authlib
    #   mlflow
cycler==0.12.1
    # via matplotlib
cyclopts==3.24.0
    # via fastmcp
databricks-sdk==0.65.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
dnspython==2.8.0
    # via email-validator
docker==7.1.0
    # via mlflow
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22.2
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.0
    # via
    #   anyio
    #   fastmcp
fastapi==0.117.1
    # via mlflow-skinny
fastmcp==2.12.3
    # via mlflow
flask==3.1.2
    # via mlflow
fonttools==4.60.0
    # via matplotlib
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via mlflow-skinny
google-auth==2.40.3
    # via databricks-sdk
graphene==3.4.3
    # via mlflow
graphql-core==3.2.6
    # via
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
gunicorn==23.0.0
    # via mlflow
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   mcp
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via
    #   mlflow-skinny
    #   opentelemetry-api
isodate==0.7.2
    # via openapi-core
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
joblib==1.5.2
    # via scikit-learn
jsonschema==4.25.1
    # via
    #   mcp
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via
    #   openapi-core
    #   openapi-spec-validator
jsonschema-specifications==2025.9.1
    # via
    #   jsonschema
    #   openapi-schema-validator
kiwisolver==1.4.9
    # via matplotlib
lazy-object-proxy==1.12.0
    # via openapi-spec-validator
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.10.6
    # via
    #   nba-model (pyproject.toml)
    #   mlflow
mcp==1.14.1
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
mlflow==3.4.0
    # via nba-model (pyproject.toml)
mlflow-skinny==3.4.0
    # via mlflow
mlflow-tracing==3.4.0
    # via mlflow
more-itertools==10.8.0
    # via openapi-core
numpy==2.2.6
    # via
    #   nba-model (pyproject.toml)
    #   contourpy
    #   matplotlib
    #   mlflow
    #   pandas
    #   scikit-learn
    #   scipy
openapi-core==0.19.5
    # via fastmcp
openapi-pydantic==0.5.1
    # via fastmcp
openapi-schema-validator==0.6.3
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.2
    # via openapi-core
opentelemetry-api==1.37.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-proto==1.37.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-sdk==1.37.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-semantic-conventions==0.58b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   gunicorn
    #   matplotlib
    #   mlflow-skinny
    #   mlflow-tracing
pandas==2.3.2
    # via
    #   nba-model (pyproject.toml)
    #   mlflow
parse==1.20.2
    # via openapi-core
pathable==0.4.4
    # via jsonschema-path
pillow==11.3.0
    # via matplotlib
protobuf==6.32.1
    # via
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-proto
pyarrow==21.0.0
    # via mlflow
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.11.9
    # via
    #   fastapi
    #   fastmcp
    #   mcp
    #   mlflow-skinny
    #   mlflow-tracing
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pygments==2.19.2
    # via rich
pyparsing==3.2.4
    # via matplotlib
pyperclip==1.10.0
    # via fastmcp
python-dateutil==2.9.0.post0
    # via
    #   graphene
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via
    #   fastmcp
    #   mlflow-skinny
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   nba-model (pyproject.toml)
    #   jsonschema-path
    #   mlflow-skinny
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   databricks-sdk
    #   docker
    #   jsonschema-path
    #   mlflow-skinny
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==14.1.0
    # via
    #   cyclopts
    #   fastmcp
    #   rich-rst
rich-rst==1.3.1
    # via cyclopts
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
scikit-learn==1.7.2
    # via
    #   nba-model (pyproject.toml)
    #   mlflow
scipy==1.15.3
    # via
    #   mlflow
    #   scikit-learn
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   alembic
    #   mlflow
sqlparse==0.5.3
    # via mlflow-skinny
sse-starlette==3.0.2
    # via mcp
starlette==0.48.0
    # via
    #   fastapi
    #   mcp
threadpoolctl==3.6.0
    # via scikit-learn
tomli==2.2.1
    # via alembic
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   cyclopts
    #   exceptiongroup
    #   fastapi
    #   graphene
    #   mlflow-skinny
    #   openapi-core
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   docker
    #   requests
uvicorn==0.36.0
    # via
    #   mcp
    #   mlflow-skinny
werkzeug==3.1.1
    # via
    #   flask
    #   openapi-core
zipp==3.23.0
    # via importlib-metadata
